Make the long amount and capacity abstract

+++ src/main/java/net/neoforged/neoforge/transfer/handlers/templates/resources/EmptyResourceHandler.java
@@ -58,12 +58,12 @@ public final class EmptyResourceHandler<T extends IResource> implements IResourc
     }
 
     @Override
-    public int getAmount(int index) {
+    public long getAmount(int index) {
         throw new IndexOutOfBoundsException(index);
     }
 
     @Override
-    public int getCapacity(int index, T resource) {
+    public long getCapacity(int index, T resource) {
         throw new IndexOutOfBoundsException(index);
     }
